+++ unix/Xvnc/programs/Xserver/hw/vnc/init.c
@@ -417,6 +417,16 @@ int ddxProcessArgument(int argc, char *argv[], int i)
     return 1;
   }
 
+  if (strcasecmp(argv[i], "-pointerlocktimeout") == 0) {
+    REQUIRE_ARG();
+    rfbPointerLockTimeout = atoi(argv[i + 1]);
+    if (rfbPointerLockTimeout < 0) {
+      UseMsg();
+      exit(1);
+    }
+    return 2;
+  }
+
   /* Run server in view-only mode - Ehud Karni SW */
   if (strcasecmp(argv[i], "-viewonly") == 0) {
     rfbViewOnly = TRUE;
@@ -1627,6 +1637,7 @@ void ddxGiveUp(enum ExitCode error)
     rfbPAMEnd(cl);
 #endif
   ShutdownTightThreads();
+  TimerFree(pointerLockTimer);
   free(rfbFB.pfbMemory);
   if (initOutputCalled) {
     char unixSocketName[32];
@@ -1738,6 +1749,12 @@ void ddxUseMsg(void)
   ErrorF("======================\n");
   ErrorF("-compatiblekbd         set META key = ALT key as in the original VNC\n");
   ErrorF("-nocursor              don't display a cursor\n");
+  ErrorF("-pointerlocktimeout time\n");
+  ErrorF("                       max time in ms (0 = indefinitely) to wait for a new\n");
+  ErrorF("                       pointer event from a connected viewer that is dragging\n");
+  ErrorF("                       the mouse (and thus has exclusive control over the\n");
+  ErrorF("                       pointer) [default: %d]\n",
+         DEFAULT_POINTER_LOCK_TIMEOUT);
   ErrorF("-viewonly              only let viewers view, not control, the remote desktop\n");
   ErrorF("-virtualtablet         set up virtual stylus and eraser devices for this\n");
   ErrorF("                       session, to emulate a Wacom tablet, and map all\n");